+++ header/lexer.h
@@ -22,6 +22,10 @@ bool lexer(t_condition *condition, char *line);
 t_token *tokenizer(char *line);
 t_token *new_token(char *token, t_token_kind kind);
 void expand_token(t_condition *condition, t_token *tokenized);
+void get_env_name(t_condition *condition, t_token *tokenized, t_lexer *info, int *i);
+void replace_env(t_token *token, char *env_value, int env_len);
+char *find_env(t_condition *condition, char *env_key);
+void append_char(char **str, char c);
