+++ src/lexer/expand.c
@@ -2,113 +2,12 @@
 #include "../../header/lexer.h"
 #include "../../libft/libft.h"
 
-void append_char(char **str, char c)
-{
-	char *new;
-	char join[2];
-
-	join[0] = c;
-	join[1] = '\0';
-	new = ft_strjoin(*str, join);
-	if(!new)
-	{
-		printf("TODO:append_char malloc failed\n");
-		exit(1);
-	}
-	free(*str);
-	*str = new;
-}
-
-void replace_env(t_token *token, char *env_value, int env_len)
-{
-	char *new;
-	int i;
-
-	if (!env_value)
-		env_value = "";//環境変数が存在しない場合要確認
-	i = 0;
-	new = ft_strdup("");
-	while(token->token[i])
-	{
-		if (token->token[i] == '$')
-		{
-			new = ft_strjoin(new, env_value);
-			i += env_len + 1;
-			new = ft_strjoin(new, token->token + i);
-			break ;
-		}
-		else
-			append_char(&new, token->token[i++]);
-	}
-	free(token->token); 
-	token->token = new;
-}
-	
-char *find_env(t_condition *condition, char *env_key)
-{
-	t_item *temp;
-
-	temp = condition->environ;
-	if (ft_strncmp(env_key, "?\0", 2) == 0)
-		return (ft_itoa(condition->exit_status));
-	while(temp)
-	{
-		if(ft_strncmp(temp->key, env_key, ft_strlen(temp->key) + 1) == 0)
-			return (temp->value);
-		temp = temp->next;
-	}
-	(void)condition;
-;	return (NULL);
-}
-
 void update_quote_status(t_lexer *info, char c)
 {
 	if (info->quote == 0)
 		info->quote = c;
 	else if (info->quote == c)
 		info->quote = 0;
-}
-
-void get_env_name(t_condition *condition, t_token *tokenized, t_lexer *info, int *i)
-{
-	int env_len;
-
-	env_len = 0;
-	(*i)++;
-	while (ft_isalnum(tokenized->token[*i + env_len]))
-		env_len++;
-	info->env_key = ft_substr(tokenized->token, *i, env_len);
-	if (!info->env_key)
-	{
-		printf("TODO:get_env_name malloc failed\n");
-		exit(1);
-	}
-	if (info->env_key[0] == '\0')
-	{
-		(*i)++;
-		return ;
-	}
-	replace_env(tokenized, find_env(condition, info->env_key), ft_strlen(info->env_key));
-}
-
-void expand_dollar(t_condition *condition, t_token *tokenized)
-{
-	int i;
-	int env_len;
-	t_lexer info;
-
-	i = 0;
-	env_len = 0;
-	ft_memset(&info, 0, sizeof(t_lexer));
-	while (tokenized->kind == TOKEN_WORD && tokenized->token[i])
-	{
-		if (is_quote(tokenized->token[i]))
-			update_quote_status(&info, tokenized->token[i]);
-		if (tokenized->token[i] == '$' && info.quote != SINGLE_QUOTE)
-			get_env_name(condition, tokenized, &info, &i);
-		else
-			i++;
-	}
 }
 
 void expand_quote(t_token *tokenized)
@@ -134,6 +33,26 @@ void expand_quote(t_token *tokenized)
 	}
 	free(tokenized->token);
 	tokenized->token = new;
+}
+
+void expand_dollar(t_condition *condition, t_token *tokenized)
+{
+	int i;
+	int env_len;
+	t_lexer info;
+
+	i = 0;
+	env_len = 0;
+	ft_memset(&info, 0, sizeof(t_lexer));
+	while (tokenized->kind == TOKEN_WORD && tokenized->token[i])
+	{
+		if (is_quote(tokenized->token[i]))
+			update_quote_status(&info, tokenized->token[i]);
+		if (tokenized->token[i] == '$' && info.quote != SINGLE_QUOTE)
+			get_env_name(condition, tokenized, &info, &i);
+		else
+			i++;
+	}
 }
 
 void expand_token(t_condition *condition, t_token *tokenized)
